Experimental refactor of tests.
Given that LLMs make strange errors like (as we're now seeing)
returning negative numbers in guidance context which they would
never make in free response context, I think we need a way of
formatting tests that makes it clear which specific test cases
are failing and which aren't.

The code of tests looks more elegant when it's formatted like
those in `tests/test_gpts_coarse.py`.

But we can see more fine-grained failures when code is formatted
like it is in `tests/test_gpts_fine.py`, at the cost of more
verbose and less elegant code.

A third approach is to have a file like `tests/eval_gpts.py`
that just returns values that we'll later assert to be equal,
so we can see more clearly what failed and how, which would
make debugging failures easier.

Sending all of the above so you can decide which you prefer.